# api/openapi.yaml
openapi: 3.0.3
info:
  title: Poydem API
  version: 1.0.0
  license:
    name: Proprietary
  description: |
    Final frontend/backend contract for Poydem MVP API v1.

    All enum values use lower-case except stable error and system-reason codes,
    which use UPPER_SNAKE_CASE. Every JSON response uses UTF-8. Every response
    includes X-Request-ID. Timestamps are RFC 3339 date-time values with an offset.

    Access tokens are sent as Bearer JWTs. Refresh tokens exist only in an HttpOnly
    cookie and are never returned in JSON or URLs. Lists documented with pagination
    return `items: []` and `totalPages: 0` when empty. Unexpected server failures
    return HTTP 500 with error.code `INTERNAL_ERROR` in the common error envelope.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /uploads/events:
    post:
      operationId: uploadEventCover
      tags: [Events]
      summary: Upload a temporary event cover
      description: The returned URL may be used by the same user in POST or PATCH event. Unlinked uploads expire after 24 hours.
      x-error-codes:
        '400': [VALIDATION_ERROR]
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, PROFILE_INCOMPLETE]
        '413': [EVENT_COVER_TOO_LARGE]
        '415': [UNSUPPORTED_EVENT_COVER_TYPE]
        '429': [RATE_LIMITED]
        '503': [EVENT_COVER_UPLOAD_FAILED]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary, description: JPEG, PNG or WebP up to 10 MiB}
      responses:
        '201':
          description: Temporary cover uploaded
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EventCoverUpload'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '413': {description: File is too large, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '415': {description: Unsupported image type, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '429': {description: Upload rate limit exceeded, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '503': {description: Object storage is unavailable, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
  /reports:
    post:
      operationId: createReport
      tags: [Reports]
      summary: Create report
      x-error-codes:
        '400': [VALIDATION_ERROR]
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, PROFILE_INCOMPLETE]
        '404': [REPORT_TARGET_NOT_FOUND]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReportInput'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Report'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /admin/reports:
    get:
      operationId: listAdminReports
      tags: [Admin]
      summary: Admin reports
      x-error-codes:
        '400': [VALIDATION_ERROR]
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, FORBIDDEN]
      parameters:
        - {name: status, in: query, schema: {type: string, enum: [pending, resolved, rejected]}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Reports
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReportList'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /admin/reports/{reportId}:
    get:
      operationId: getAdminReport
      tags: [Admin]
      summary: Admin report details
      x-error-codes:
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, FORBIDDEN]
        '404': [REPORT_NOT_FOUND]
      parameters:
        - {name: reportId, in: path, required: true, schema: {type: integer, format: int64}}
      responses:
        '200':
          description: Report
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Report'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /admin/reports/{reportId}/{action}:
    post:
      operationId: moderateReport
      tags: [Admin]
      summary: Resolve or reject report
      x-error-codes:
        '401': [UNAUTHORIZED]
        '403': [USER_BANNED, FORBIDDEN]
        '404': [REPORT_NOT_FOUND]
        '409': [REPORT_ALREADY_RESOLVED]
      parameters:
        - {name: reportId, in: path, required: true, schema: {type: integer, format: int64}}
        - {name: action, in: path, required: true, schema: {type: string, enum: [resolve, reject]}}
      responses:
        '200':
          description: Processed
          content: {application/json: {schema: {$ref: '#/components/schemas/Report'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly cookie managed by the browser; JavaScript cannot read it
  parameters:
...
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
  responses:
    BadRequest:
      description: Bad request
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Unauthorized:
      description: Unauthorized
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Forbidden:
      description: Forbidden
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotFound:
      description: Not found
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Conflict:
      description: Conflict
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
  schemas:
...
    UserShort:
      type: object
      required: [id, firstName, lastName, avatarUrl]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        firstName: {type: string}
        lastName: {type: string, nullable: true}
        avatarUrl: {type: string, format: uri, nullable: true}
...
    ReportInput:
      type: object
      required: [targetType, targetId, reason]
      properties:
        targetType: {type: string, enum: [user, company, event]}
        targetId: {type: integer, format: int64, minimum: 1}
        reason: {type: string, minLength: 1, maxLength: 100}
        description: {type: string, maxLength: 2000, nullable: true}
    Report:
      allOf:
        - {$ref: '#/components/schemas/ReportInput'}
        - type: object
          required: [id, description, author, status, resolvedBy, createdAt, resolvedAt]
          properties:
            id: {type: integer, format: int64, minimum: 1}
            author: {$ref: '#/components/schemas/UserShort'}
            status: {type: string, enum: [pending, resolved, rejected]}
            resolvedBy:
              type: object
              nullable: true
              allOf: [{$ref: '#/components/schemas/UserShort'}]
            createdAt: {type: string, format: date-time}
            resolvedAt: {type: string, format: date-time, nullable: true}
    ReportList:
      type: object
      required: [items, pagination]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Report'}}
        pagination: {$ref: '#/components/schemas/Pagination'}
...
    EventCoverUpload:
      type: object
      required: [url]
      additionalProperties: false
      properties:
        url: {type: string, format: uri}
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: {type: integer, minimum: 1}
        limit: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
        totalPages: {type: integer, minimum: 0, description: Zero when total is zero}
    ErrorResponse:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, details]
          additionalProperties: false
          properties:
            code: {$ref: '#/components/schemas/ErrorCode'}
            message: {type: string, example: No free places in company}
            details: {$ref: '#/components/schemas/ErrorDetails'}
    ErrorDetails:
      type: object
      properties:
        fields:
          type: object
          additionalProperties:
            type: array
            items: {type: string}
          description: Present for VALIDATION_ERROR; keys are request field or parameter names
      additionalProperties: true
    ErrorCode:
      type: string
      description: Stable machine-readable code. Frontend logic must never branch on error.message.
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - INTERNAL_ERROR
        - OAUTH_PROVIDER_UNAVAILABLE
        - USER_BANNED
        - USER_NOT_FOUND
        - PROFILE_INCOMPLETE
        - EVENT_NOT_FOUND
        - EVENT_NOT_AVAILABLE
        - INVALID_STATUS_TRANSITION
        - COMPANY_NOT_FOUND
        - COMPANY_FULL
        - COMPANY_CLOSED
        - COMPANY_BLOCKED
        - COMPANY_CAPACITY_BELOW_MEMBERS
        - NOT_COMPANY_OWNER
        - NOT_COMPANY_MEMBER
        - OWNER_CANNOT_LEAVE
        - OWNER_CANNOT_BE_REMOVED
        - ALREADY_COMPANY_MEMBER
        - ALREADY_IN_EVENT_COMPANY
        - APPLICATION_ALREADY_EXISTS
        - APPLICATION_NOT_FOUND
        - APPLICATION_ALREADY_RESOLVED
        - ALREADY_EVENT_PARTICIPANT
        - NOT_SOLO_PARTICIPANT
        - REPORT_TARGET_NOT_FOUND
        - REPORT_NOT_FOUND
        - REPORT_ALREADY_RESOLVED
        - AVATAR_TOO_LARGE
        - UNSUPPORTED_AVATAR_TYPE
        - AVATAR_UPLOAD_FAILED
        - EVENT_FORBIDDEN
        - EVENT_NOT_EDITABLE
        - EVENT_COVER_TOO_LARGE
        - UNSUPPORTED_EVENT_COVER_TYPE
        - EVENT_COVER_UPLOAD_FAILED
        - RATE_LIMITED
        - APPLICANT_BANNED
        - CANNOT_CHANGE_OWN_STATUS
        - CANNOT_BAN_ADMIN